Move conversion checker to utility
