+++ src/xunit.analyzers/X1000/InlineDataMustMatchTheoryParameters.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
@@ -6,6 +5,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
+using Xunit.Analyzers.Utility;
 
 namespace Xunit.Analyzers;
 
@@ -240,78 +240,5 @@ public class InlineDataMustMatchTheoryParameters : XunitDiagnosticAnalyzer
 		/// E.g. InlineData(data: new object[] { 1, 2, 3 }) or InlineData(new object[] { 1, 2, 3 })
 		/// </summary>
 		Initializer,
-	}
-
-	static class ConversionChecker
-	{
-		public static bool IsConvertible(
-			Compilation compilation,
-			ITypeSymbol source,
-			ITypeSymbol destination,
-			XunitContext xunitContext)
-		{
-			if (destination.TypeKind == TypeKind.Array)
-			{
-				var destinationElementType = ((IArrayTypeSymbol)destination).ElementType;
-
-				if (destinationElementType.TypeKind == TypeKind.TypeParameter)
-					return IsConvertibleTypeParameter(source, (ITypeParameterSymbol)destinationElementType);
-			}
-
-			if (destination.TypeKind == TypeKind.TypeParameter)
-				return IsConvertibleTypeParameter(source, (ITypeParameterSymbol)destination);
-
-			var conversion = compilation.ClassifyConversion(source, destination);
-
-			if (conversion.IsNumeric)
-				return IsConvertibleNumeric(source, destination);
-
-			if (destination.SpecialType == SpecialType.System_DateTime
-				|| (xunitContext.Core.TheorySupportsConversionFromStringToDateTimeOffsetAndGuid == true && IsDateTimeOffsetOrGuid(destination)))
-			{
-				// Allow all conversions from strings. All parsing issues will be reported at runtime.
-				return source.SpecialType == SpecialType.System_String;
-			}
-
-			// Rules of last resort
-			return conversion.IsImplicit
-				|| conversion.IsUnboxing
-				|| (conversion.IsExplicit && conversion.IsUserDefined)
-				|| (conversion.IsExplicit && conversion.IsNullable);
-		}
-
-		static bool IsConvertibleTypeParameter(
-			ITypeSymbol source,
-			ITypeParameterSymbol destination)
-		{
-			if (destination.HasValueTypeConstraint && !source.IsValueType)
-				return false;
-			if (destination.HasReferenceTypeConstraint && source.IsValueType)
-				return false;
-
-			return destination.ConstraintTypes.All(c => c.IsAssignableFrom(source));
-		}
-
-		static bool IsConvertibleNumeric(
-			ITypeSymbol source,
-			ITypeSymbol destination)
-		{
-			if (destination.SpecialType == SpecialType.System_Char
-				&& (source.SpecialType == SpecialType.System_Double || source.SpecialType == SpecialType.System_Single))
-			{
-				// Conversions from float to char (though numeric) do not actually work at runtime, so report them
-				return false;
-			}
-
-			return true; // Allow all numeric conversions. Narrowing conversion issues will be reported at runtime.
-		}
-
-		static bool IsDateTimeOffsetOrGuid(ITypeSymbol destination)
-		{
-			if (destination.ContainingNamespace?.Name != nameof(System))
-				return false;
-
-			return destination.MetadataName == nameof(DateTimeOffset) || destination.MetadataName == nameof(Guid);
-		}
 	}
 }
